Bl 841 reindex skip supressed

+++ scripts/ingest_marc.sh
@@ -9,9 +9,15 @@ elif [[ -s "/usr/local/rvm/scripts/rvm" ]] ; then
 else
   printf "ERROR: An RVM installation was not found.\n"
 fi
+
+if [ ! -e ${1} ]
+then
+  exit 1
+fi
+
 cd $HOME/tul_cob
 rvm use 2.4.1@tul_cob
 gem install bundler
 bundle install
-bundle exec traject -c lib/traject/indexer_config.rb ${1}
+SOLR_URL="$2" bundle exec traject -c lib/traject/indexer_config.rb ${1}
 return 0
